langgraph_agent_template/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langgraph
langchain-google-genai
pydantic>=2.0.0
a2a-sdk[http-server]
